Update to flutter 2 with null safety
