+++ localization_demo/pubspec.yaml
@@ -6,7 +6,7 @@ publish_to: "none"
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -15,10 +15,10 @@ dependencies:
   cupertino_icons: ^0.1.3
 
   # Highlighted package
-  easy_localization: ^2.3.2
+  easy_localization: ^3.0.0
 
   # For UI template
-  velocity_x: ^0.3.2
+  velocity_x: ^3.0.0
 
 dev_dependencies:
   flutter_test:
